fix parameter types from double to float

+++ tracer/src/Datadog.Trace/TraceContext.cs
@@ -93,9 +93,9 @@ namespace Datadog.Trace
 
         public string? SamplingMechanism { get; set; }
 
-        public double? AppliedSamplingRate { get; set; }
+        public float? AppliedSamplingRate { get; set; }
 
-        public double? RateLimiterRate { get; set; }
+        public float? RateLimiterRate { get; set; }
 
         public double? TracesKeepRate { get; set; }
 
@@ -297,8 +297,8 @@ namespace Datadog.Trace
         public void SetSamplingPriority(
             int? priority,
             string? mechanism = null,
-            double? rate = null,
-            double? limiterRate = null,
+            float? rate = null,
+            float? limiterRate = null,
             bool notifyDistributedTracer = true)
         {
             if (priority is not { } p)
